crud update on personal,employement, fileupload liability assets & extra income

+++ api/objects/ha_customer_assets.php
@@ -332,7 +332,7 @@ class Ha_Customer_Assets
     {
 
         // delete query
-        $query = "DELETE FROM " . $this->table_name . " WHERE id = ? ";
+        $query = "DELETE FROM " . $this->table_name . " WHERE id = ? AND user_id = ? ";
 
         // prepare query
         $stmt = $this->conn->prepare($query);
@@ -342,6 +342,7 @@ class Ha_Customer_Assets
 
         // bind id of record to delete
         $stmt->bindParam(1, $this->id);
+        $stmt->bindParam(2, $this->user_id);
         $stmt->execute();
 
         if ($stmt->rowCount()) {
